(at last) push month-old itemwise analysis
see itemwise_analysis folder, seems to suggest structure of the sentences is responsible for our results and there's little variance between words.
The fact that words that only appear as fillers or as gaps have such similar stats to words that always appear suggests (at least to me) that the effect of the use of a word like "shoe" vs using "snack" vs using "bag" is minor, and well-controlled–for in experiment's design.